Get new git changes to build series
Also had to mess around with the scraper to slow down the requests. I dont know what the limit is, but trying to download more than one series will get us temporarily blocked.

+++ scripts/SetImport/ImageScrape/ScrapeV2.js
@@ -7,53 +7,63 @@ var IMAGE_DESTINATION = __dirname +'/../../../public/images/JP'
 
 var EXTENSION = process.env.IMAGE_EXTENSION || 'png';
 var APPEND_TRIAL = false;//set to false for old sets that dont need -td appended
+const IMAGE_DOWNLOAD_RATELIMIT = 200;//Requesting to many images quickly will get us 403'd.
 
 var download = (url, filename) => {
 
   return new Promise(resolve => {
     axios({method: "get", url, responseType: 'stream'})
     .then((res) => {
-      res.data.pipe(createWriteStream(filename)).on('close', resolve);
+      setTimeout(() => {
+        res.data.pipe(createWriteStream(filename)).on('close', resolve);
+      }, IMAGE_DOWNLOAD_RATELIMIT)      
+    })
+    .catch((err) => {
+      console.log('Request error', url, err.response)
+        resolve()      
     })
   }) 
-
-  // request.head(uri, function(err, res, body){
-  //   console.log('content-type:', res.headers['content-type']);
-  //   console.log('content-length:', res.headers['content-length']);
-
-  //   request(uri).pipe(createWriteStream(filename)).on('close', callback);
-  // });
 };
 
-const Scrape = async(SET_FILE) => {
+const Scrape = (SET_FILE) => {
 
   console.log('Scraping images...', SET_FILE)
 
   let set = JSON.parse(readFileSync(`${SET_DATA_PATH}/${SET_FILE}.json`, { encoding: 'utf8'}));
 
-  for (const card of set) {
-    var trial =  APPEND_TRIAL && card.sid.indexOf('T') !== -1;
+  return new Promise(async(resolve, reject) => {
+    for (const card of set) {
+      var trial =  APPEND_TRIAL && card.sid.indexOf('T') !== -1;
 
-    var cardset = `${card.set.toLowerCase()}_${card.side.toLowerCase()}${card.release.toLowerCase()}`
-    var cardname = `${cardset}_${card.sid.toLowerCase()}.${EXTENSION}`
+      var cardset = `${card.set.toLowerCase()}_${card.side.toLowerCase()}${card.release.toLowerCase()}`
+      var cardname = `${cardset}_${card.sid.toLowerCase()}.${EXTENSION}`
 
-    var remotepath = `https://ws-tcg.com/wordpress/wp-content/images/cardlist/${card.set.toLowerCase().charAt(0)}/${card.set.toLowerCase()}_${card.side.toLowerCase()}${card.release.toLowerCase()}/${cardname}`;
-    console.log(remotepath)
+      var remotepath = `https://ws-tcg.com/wordpress/wp-content/images/cardlist/${card.set.toLowerCase().charAt(0)}/${card.set.toLowerCase()}_${card.side.toLowerCase()}${card.release.toLowerCase()}/${cardname}`;
+      console.log(remotepath)
 
-    !existsSync(`${IMAGE_DESTINATION}/${card.side}${card.release}`) && mkdirSync(`${IMAGE_DESTINATION}/${card.side}${card.release}`);
+      !existsSync(`${IMAGE_DESTINATION}/${card.side}${card.release}`) && mkdirSync(`${IMAGE_DESTINATION}/${card.side}${card.release}`);
 
-     await download(remotepath, `${IMAGE_DESTINATION}/${card.side}${card.release}/${card.sid}.${EXTENSION}`)
-     .then(() => {
-        if(EXTENSION != 'gif'){
-          sharp(`${IMAGE_DESTINATION}/${card.side}${card.release}/${card.sid}.${EXTENSION}`)
-          .rotate()
-          .toFile(`${IMAGE_DESTINATION}/${card.side}${card.release}/${card.sid}.gif`, (err, info) => {
-            console.log('Converted to gif')
-            unlink(`${IMAGE_DESTINATION}/${card.side}${card.release}/${card.sid}.${EXTENSION}`, (err) => err ? console.log(err) : console.log('Done'))
-          })
-        }    
-     })
-  }
+      //If the .gif image already exists, skip downloading
+      if( !existsSync(`${IMAGE_DESTINATION}/${card.side}${card.release}/${card.sid}.gif`) ){
+        await download(remotepath, `${IMAGE_DESTINATION}/${card.side}${card.release}/${card.sid}.${EXTENSION}`)
+        .then(async() => {
+          if(EXTENSION != 'gif'){
+            await sharp(`${IMAGE_DESTINATION}/${card.side}${card.release}/${card.sid}.${EXTENSION}`)
+            .rotate()
+            .toFile(`${IMAGE_DESTINATION}/${card.side}${card.release}/${card.sid}.gif`, (err, info) => {
+              console.log('Converted to gif')
+              unlink(`${IMAGE_DESTINATION}/${card.side}${card.release}/${card.sid}.${EXTENSION}`, (err) => err ? console.log(err) : console.log('Done'))
+            })
+          }    
+        })
+      }
+      else{
+        console.log('Skipping download: ' + `${IMAGE_DESTINATION}/${card.side}${card.release}/${card.sid}.${EXTENSION}`)
+      }
+
+    }
+    resolve()
+  })
 }
 
 module.exports = Scrape;
